requirements.txt
praw>=7.7.0
pandas>=1.3.0
streamlit>=1.3.0
plotly>=5.5.0
matplotlib>=3.5.0
python-dotenv>=0.20.0
